relocate menuData to mock directory as more mock data will be introduced
